Add safety check for max memory, load as f16
The maximum available memory and swap are checked before loading a video
to ensure sufficient space exists to both load, and process the video
data.

Note that the load is attempted even if frame_load_cap is set to above
the calculated memory limit in case the actual video has fewer frames
than this cap.

When loading videos and images, the images are loaded as f16 instead of
f32. This results in a ~25% memory reduction for most workflows and
should ensure the slight overhead of the above safety checks does not
create any regression in what workflows are capable of being executed.
This change has been fairly extensive testing on many vanilla nodes, but
has a fairly high potential to cause issues and should be watched
carefully.

+++ videohelpersuite/load_images_nodes.py
@@ -61,7 +61,7 @@ def images_generator(directory: str, image_load_cap: int = 0, skip_first_images:
         i = Image.open(file_path)
         i = ImageOps.exif_transpose(i)
         i = i.convert(iformat)
-        i = np.array(i, dtype=np.float32) / 255.0
+        i = np.array(i, dtype=np.float16) / 255.0
         if has_alpha:
             i[:,:,-1] = 1 - i[:,:,-1]
         if i.shape[0] != size[1] or i.shape[1] != size[0]:
@@ -97,13 +97,13 @@ def load_images(directory: str, image_load_cap: int = 0, skip_first_images: int 
 
     if meta_batch is not None:
         gen = itertools.islice(gen, meta_batch.frames_per_batch)
-    images = torch.from_numpy(np.fromiter(gen, np.dtype((np.float32, (height, width, 3 + has_alpha)))))
+    images = torch.from_numpy(np.fromiter(gen, np.dtype((np.float16, (height, width, 3 + has_alpha)))))
     if has_alpha:
         #tensors are not continuous. Rewrite will be required if this is an issue
         masks = images[:,:,:,3:]
         images = images[:,:,:,:3]
     else:
-        masks = torch.zeros((images.size(0), 64, 64), dtype=torch.float32, device="cpu")
+        masks = torch.zeros((images.size(0), 64, 64), dtype=torch.float16, device="cpu")
     if len(images) == 0:
         raise FileNotFoundError(f"No images could be loaded from directory '{directory}'.")
